refactor: adjust some functions's position
