+++ src/geometry/animate/util.ts
@@ -1,68 +1,6 @@
 import * as _ from '@antv/util';
 import { Coordinate, Shape } from '../../dependents';
-import { Point, ShapeDrawCFG } from '../../interface';
-
-// 获取图形的包围盒
-function getPointsBox(points) {
-  if (_.isEmpty(points)) {
-    return null;
-  }
-
-  let minX = points[0].x;
-  let maxX = points[0].x;
-  let minY = points[0].y;
-  let maxY = points[0].y;
-  _.each(points, (point) => {
-    minX = minX > point.x ? point.x : minX;
-    maxX = maxX < point.x ? point.x : maxX;
-    minY = minY > point.y ? point.y : minY;
-    maxY = maxY < point.y ? point.y : maxY;
-  });
-
-  return {
-    minX,
-    maxX,
-    minY,
-    maxY,
-    centerX: (minX + maxX) / 2,
-    centerY: (minY + maxY) / 2,
-  };
-}
-
-export function polarToCartesian(centerX: number, centerY: number, radius: number, angleInRadian: number) {
-  return {
-    x: centerX + radius * Math.cos(angleInRadian),
-    y: centerY + radius * Math.sin(angleInRadian),
-  };
-}
-
-export function getSectorPath(
-  centerX: number,
-  centerY: number,
-  radius: number,
-  startAngleInRadian: number,
-  endAngleInRadian: number
-) {
-  const start = polarToCartesian(centerX, centerY, radius, endAngleInRadian);
-  const end = polarToCartesian(centerX, centerY, radius, startAngleInRadian);
-
-  if (endAngleInRadian - startAngleInRadian === Math.PI * 2) {
-    return [
-      ['M', centerX, centerY],
-      ['m', -radius, 0],
-      ['a', radius, radius, 0, 1, 0, radius * 2, 0],
-      ['a', radius, radius, 0, 1, 0, -radius * 2, 0],
-    ];
-  }
-
-  const arcSweep = endAngleInRadian - startAngleInRadian <= Math.PI ? 0 : 1;
-  return [
-    ['M', start.x, start.y],
-    ['A', radius, radius, 0, arcSweep, 0, end.x, end.y],
-    ['L', centerX, centerY],
-    ['L', start.x, start.y],
-  ];
-}
+import { getSectorPath } from '../../util/graphics';
 
 export function getCoordinateClipCfg(coordinate: Coordinate) {
   const { start, end } = coordinate;
@@ -120,28 +58,5 @@ export function getCoordinateClipCfg(coordinate: Coordinate) {
       height: coordinate.isTransposed ? 0 : height + margin * 2,
     },
     endState,
-  };
-}
-
-export function getAngle(shapeModel: ShapeDrawCFG, coordinate: Coordinate) {
-  const points = shapeModel.points;
-  const box = getPointsBox(points);
-  let endAngle;
-  let startAngle;
-  const { startAngle: coordStartAngle, endAngle: coordEndAngle } = coordinate;
-  const diffAngle = coordEndAngle - coordStartAngle;
-
-  if (coordinate.isTransposed) {
-    endAngle = box.maxY * diffAngle;
-    startAngle = box.minY * diffAngle;
-  } else {
-    endAngle = box.maxX * diffAngle;
-    startAngle = box.minX * diffAngle;
-  }
-  endAngle += coordStartAngle;
-  startAngle += coordStartAngle;
-  return {
-    startAngle,
-    endAngle,
   };
 }
